if anybody wants he can code it i don't know how to

+++ Src/bldc.c
@@ -230,7 +230,7 @@ void sensored_brushless_countrol(){
   }
 }
 uint8_t nextPos(uint8_t oldpos, bool forward){
-  return (forward ? oldpos + 1 : oldpos - 1) & 0x7;
+  return ((uint8_t)(forward ? oldpos + 1 : oldpos - 1) & 0x7);
 }
 
 void sensorless_brushless_countrol(){  // TODO: Only currentdriven control because the suckers doesnt have the voltage sensors
@@ -245,7 +245,7 @@ void sensorless_brushless_countrol(){  // TODO: Only currentdriven control becau
       adc_array[4-2*x] - adc_offset[4-2*x],
       &tmp_phase_current);  // Block Phase current for sensorless bldc control
     set_tim_lr[x]((currentlr[x] = ABS(adc_array[5-x] - adc_offset[5-x]) * MOTOR_AMP_CONV_DC_AMP) <= current_limit);
-    if(tmp_phase_current < blockcurlr[x])
+    if(tmp_phase_current < blockcurlr[x])  // check for an phase change TODO need an goot algorythom
       uint8_t pos = nextPos(last_pos[x], throttlelr > 0);  // needs to be calced
     RetValWeak tmp = currentWeaking(throttlelr[x], phase_period[x],timer[x],currentlr[x]);
     blockPWM(tmp.pwm, pos, &phase[0], &phase[1], &phase[2]);
